Lekha/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-treebeard==4.5.1
Pillow==9.2.0
sqlparse==0.4.2
